CASSANDRA-19369 Use XXHash32 for digest calculation of SSTables
This commit adds the ability to use the newly supported in Cassandra Sidecar XXhash32 digest algorithm.
The commit allows for backwards compatibility to perform MD5 checksumming, but it now defaults to XXHash32.

A new Writer option is added:

```
.option(WriterOptions.DIGEST.name(), "XXHASH32") // or
.option(WriterOptions.DIGEST.name(), "MD5")
```

This option defaults to XXHash32, when not provided, but it can be configured to use the legacy MD5 algorithm.

Path by Francisco Guerrero; Reviewed by Yifan Cai for CASSANDRA-19369