CASSANDRA-19500 Mask hash value produced by lz4's xxhash implementation
This PR bumps the Sidecar version to the current latest HEAD of Sidecar. Bumping the
version surfaced an issue in the LZ4's implementation of the XXHash32 hash value. The
hash value is not masked and this causes the negative sign to be forwarded producing
the incorrect hash result. Other implemnentations correctly apply the mask to the
produced value (i.e commons-codec).

We do not consider switching to `commons-codec` because Spark has a dependency of
codec, and in older versions of Spark (2), the XXHash32 implementation is either non
existent or the APIs have changed.

Patch by Francisco Guerrero; Reviewed by TBD for CASSANDRA-19500