CASSANDRA-19507 Fix bulk reads of multiple tables that potentially have the same data file name
When reading multiple data frames using bulk reader from different tables, it is possible to encounter a data
file name being retrieved from the same Sidecar instance. Because the `SSTable`s are cached in the `SSTableCache`,
it is possible that the `org.apache.cassandra.spark.reader.SSTableReader` uses the incorrect `SSTable` if it was
cached with the same `#hashCode`.

In this patch, the equality takes into account the keyspace, table, and snapshot name.

Patch by Francisco Guerrero; Reviewed by TBD for CASSANDRA-19507